backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
asyncpg==0.30.0
openai==2.1.0
anthropic==0.69.0
python-dotenv==1.1.1
sentence-transformers==5.1.1
pydantic==2.11.9
numpy==2.3.3
PyPDF2==3.0.1
python-multipart==0.0.20